refactor: move Block related components
